Refactor ratelimiting #12
